refactor(api): move network.js utils from iam to shared
