+++ api/src/identity-access-management/application/oidc-provider/oidc-provider.controller.js
@@ -1,11 +1,11 @@
 import { BadRequestError, UnauthorizedError } from '../../../shared/application/errors/http-errors.js';
 import * as localeService from '../../../shared/domain/services/locale-service.js';
 import { logger } from '../../../shared/infrastructure/utils/logger.js';
+import { getForwardedOrigin, RequestedApplication } from '../../../shared/infrastructure/utils/network.js';
 import { getUserLocale } from '../../../shared/infrastructure/utils/request-response-utils.js';
 import { usecases } from '../../domain/usecases/index.js';
 import * as oidcProviderSerializer from '../../infrastructure/serializers/jsonapi/oidc-identity-providers.serializer.js';
 import * as userOidcAuthenticationRequestSerializer from '../../infrastructure/serializers/jsonapi/user-oidc-authentication-request.serializer.js';
-import { getForwardedOrigin, RequestedApplication } from '../../infrastructure/utils/network.js';
 
 /**
  * @typedef {function} authenticateOidcUser
