+++ api/src/prescription/organization-learner/application/sco-organization-learner-controller.js
@@ -1,11 +1,8 @@
 import dayjs from 'dayjs';
 
-import {
-  getForwardedOrigin,
-  RequestedApplication,
-} from '../../../identity-access-management/infrastructure/utils/network.js';
 import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 import { getI18nFromRequest } from '../../../shared/infrastructure/i18n/i18n.js';
+import { getForwardedOrigin, RequestedApplication } from '../../../shared/infrastructure/utils/network.js';
 import { getUserLocale } from '../../../shared/infrastructure/utils/request-response-utils.js';
 import * as scoOrganizationLearnerSerializer from '../../learner-management/infrastructure/serializers/jsonapi/sco-organization-learner-serializer.js';
 import { usecases } from '../domain/usecases/index.js';
