+++ api/src/shared/infrastructure/plugins/pino.js
@@ -1,10 +1,10 @@
 import { stdSerializers } from 'pino';
 
-import { getForwardedOrigin } from '../../../identity-access-management/infrastructure/utils/network.js';
 import { config } from '../../config.js';
 import { getCorrelationInfo, getInContext, setInContext } from '../execution-context-manager.js';
 import { generateHash } from '../utils/crypto.js';
 import { loggerPino } from '../utils/logger.js';
+import { getForwardedOrigin } from '../utils/network.js';
 import { routeDomainToOwnerTeam } from '../utils/route-domain-to-owner-team.js';
 
 const serializersSym = Symbol.for('pino.serializers');
