+++ api/tests/identity-access-management/unit/application/saml/saml.controller.test.js
@@ -2,7 +2,7 @@ import sinon from 'sinon';
 
 import { samlController } from '../../../../../src/identity-access-management/application/saml/saml.controller.js';
 import { usecases } from '../../../../../src/identity-access-management/domain/usecases/index.js';
-import { RequestedApplication } from '../../../../../src/identity-access-management/infrastructure/utils/network.js';
+import { RequestedApplication } from '../../../../../src/shared/infrastructure/utils/network.js';
 import { expect } from '../../../../test-helper.js';
 import { hFake } from '../../../../tooling/mocks/hapi.mock.js';
