+++ api/tests/identity-access-management/unit/application/token.controller.test.js
@@ -3,7 +3,7 @@ import sinon from 'sinon';
 import { tokenController } from '../../../../src/identity-access-management/application/token/token.controller.js';
 import { UserAccessToken } from '../../../../src/identity-access-management/domain/models/UserAccessToken.js';
 import { usecases } from '../../../../src/identity-access-management/domain/usecases/index.js';
-import { RequestedApplication } from '../../../../src/identity-access-management/infrastructure/utils/network.js';
+import { RequestedApplication } from '../../../../src/shared/infrastructure/utils/network.js';
 import { expect } from '../../../test-helper.js';
 import { hFake } from '../../../tooling/mocks/hapi.mock.js';
